requirements.txt

streamlit
speechrecognition
gtts
google-generativeai
pydub
ffmpeg
